+++ tests/placeAndMakeSwiftPerp.ts
@@ -6,7 +6,6 @@ import { Program } from '@coral-xyz/anchor';
 import {
 	ComputeBudgetProgram,
 	Keypair,
-	SendTransactionError,
 	Transaction,
 	TransactionMessage,
 	VersionedTransaction,
@@ -28,7 +27,6 @@ import {
 	loadKeypair,
 	getMarketOrderParams,
 	MarketType,
-	DriftClient,
 } from '../sdk/src';
 
 import {
@@ -102,10 +100,11 @@ describe('place and make swift order', () => {
 		spotMarketIndexes = [0, 1];
 		oracleInfos = [{ publicKey: solUsd, source: OracleSource.PYTH }];
 
+		const wallet = new Wallet(loadKeypair(process.env.ANCHOR_WALLET));
 		makerDriftClient = new TestClient({
 			connection,
 			//@ts-ignore
-			wallet: new Wallet(loadKeypair(process.env.ANCHOR_WALLET)),
+			wallet,
 			programID: chProgram.programId,
 			opts: {
 				commitment: 'confirmed',
@@ -118,6 +117,7 @@ describe('place and make swift order', () => {
 				type: 'polling',
 				accountLoader: bulkAccountLoader,
 			},
+			txVersion: 'legacy',
 		});
 		await makerDriftClient.initialize(usdcMint.publicKey, true);
 		await makerDriftClient.subscribe();
@@ -244,23 +244,27 @@ describe('place and make swift order', () => {
 			takerOrderParamsMessage
 		);
 
-		let ixs = await makerDriftClient.getPlaceAndMakeSwiftPerpOrderIxs(
-			takerDriftClient.encodeSwiftOrderParamsMessage(takerOrderParamsMessage),
-			takerOrderParamsSig,
-			uuid,
-			{
-				taker: await takerDriftClient.getUserAccountPublicKey(),
-				takerUserAccount: takerDriftClient.getUserAccount(),
-				takerStats: takerDriftClient.getUserStatsAccountPublicKey(),
-			},
-			makerOrderParams
-		);
-		ixs = [
+		const ixs = [
 			ComputeBudgetProgram.setComputeUnitLimit({
 				units: 10_000_000,
 			}),
-			...ixs,
 		];
+		ixs.push(
+			...(await makerDriftClient.getPlaceAndMakeSwiftPerpOrderIxs(
+				takerDriftClient.encodeSwiftOrderParamsMessage(takerOrderParamsMessage),
+				takerOrderParamsSig,
+				uuid,
+				{
+					taker: await takerDriftClient.getUserAccountPublicKey(),
+					takerUserAccount: takerDriftClient.getUserAccount(),
+					takerStats: takerDriftClient.getUserStatsAccountPublicKey(),
+				},
+				makerOrderParams,
+				undefined,
+				undefined,
+				ixs
+			))
+		);
 
 		const message = new TransactionMessage({
 			instructions: ixs,
@@ -278,119 +282,6 @@ describe('place and make swift order', () => {
 	});
 
 	it('should fail on bad order params sig', async () => {
-		const keypair = new Keypair();
-		await provider.connection.requestAirdrop(keypair.publicKey, 10 ** 9);
-		await sleep(1000);
-		const wallet = new Wallet(keypair);
-		const userUSDCAccount = await mockUserUSDCAccount(
-			usdcMint,
-			usdcAmount,
-			provider,
-			keypair.publicKey
-		);
-		const takerDriftClient = new DriftClient({
-			connection,
-			wallet,
-			programID: chProgram.programId,
-			opts: {
-				commitment: 'confirmed',
-			},
-			activeSubAccountId: 0,
-			perpMarketIndexes: marketIndexes,
-			spotMarketIndexes: spotMarketIndexes,
-			oracleInfos,
-			userStats: true,
-			accountSubscription: {
-				type: 'polling',
-				accountLoader: bulkAccountLoader,
-			},
-		});
-		await takerDriftClient.subscribe();
-		await takerDriftClient.initializeUserAccountAndDepositCollateral(
-			usdcAmount,
-			userUSDCAccount.publicKey
-		);
-		const takerDriftClientUser = new User({
-			driftClient: takerDriftClient,
-			userAccountPublicKey: await takerDriftClient.getUserAccountPublicKey(),
-			accountSubscription: {
-				type: 'polling',
-				accountLoader: bulkAccountLoader,
-			},
-		});
-		await takerDriftClientUser.subscribe();
-		await takerDriftClient.initializeSwiftUserOrders(
-			takerDriftClientUser.userAccountPublicKey,
-			32
-		);
-
-		const marketIndex = 0;
-		const baseAssetAmount = BASE_PRECISION;
-		const takerOrderParams = getMarketOrderParams({
-			marketIndex,
-			direction: PositionDirection.LONG,
-			baseAssetAmount,
-			price: new BN(34).mul(PRICE_PRECISION),
-			auctionStartPrice: new BN(33).mul(PRICE_PRECISION),
-			auctionEndPrice: new BN(34).mul(PRICE_PRECISION),
-			auctionDuration: 10,
-			userOrderId: 1,
-			postOnly: PostOnlyParams.NONE,
-		});
-
-		await takerDriftClientUser.fetchAccounts();
-		const makerOrderParams = getLimitOrderParams({
-			marketIndex,
-			direction: PositionDirection.SHORT,
-			baseAssetAmount,
-			price: new BN(33).mul(PRICE_PRECISION),
-			userOrderId: 1,
-			postOnly: PostOnlyParams.MUST_POST_ONLY,
-			immediateOrCancel: true,
-		});
-
-		const uuid = Uint8Array.from(Buffer.from(nanoid(8)));
-		const takerOrderParamsMessage: SwiftOrderParamsMessage = {
-			swiftOrderParams: takerOrderParams,
-			takeProfitOrderParams: null,
-			subAccountId: 0,
-			slot: new BN(await connection.getSlot()),
-			uuid,
-			stopLossOrderParams: null,
-		};
-		const takerOrderParamsSig = makerDriftClient.signSwiftOrderParamsMessage(
-			takerOrderParamsMessage
-		);
-
-		try {
-			let ixs = await makerDriftClient.getPlaceAndMakeSwiftPerpOrderIxs(
-				takerDriftClient.encodeSwiftOrderParamsMessage(takerOrderParamsMessage),
-				takerOrderParamsSig,
-				uuid,
-				{
-					taker: await takerDriftClient.getUserAccountPublicKey(),
-					takerUserAccount: takerDriftClient.getUserAccount(),
-					takerStats: takerDriftClient.getUserStatsAccountPublicKey(),
-				},
-				makerOrderParams
-			);
-			ixs = [
-				ComputeBudgetProgram.setComputeUnitLimit({
-					units: 10_000_000,
-				}),
-				...ixs,
-			];
-			await makerDriftClient.sendTransaction(new Transaction().add(...ixs));
-			assert.fail('Should have failed');
-		} catch (error) {
-			assert.equal(
-				error.transactionMessage,
-				'Transaction precompile verification failure InvalidAccountIndex'
-			);
-		}
-	});
-
-	it('should fail if diff signed message to verify ixs vs drift ixs', async () => {
 		const keypair = new Keypair();
 		await provider.connection.requestAirdrop(keypair.publicKey, 10 ** 9);
 		await sleep(1000);
@@ -453,92 +344,66 @@ describe('place and make swift order', () => {
 			postOnly: PostOnlyParams.NONE,
 			marketType: MarketType.PERP,
 		});
+		const uuid = Uint8Array.from(Buffer.from(nanoid(8)));
 		const takerOrderParamsMessage: SwiftOrderParamsMessage = {
 			swiftOrderParams: takerOrderParams,
 			subAccountId: 0,
-			uuid: Uint8Array.from(Buffer.from(nanoid(8))),
 			slot: new BN(await connection.getSlot()),
+			uuid,
 			takeProfitOrderParams: null,
 			stopLossOrderParams: null,
 		};
 
 		await takerDriftClientUser.fetchAccounts();
 
-		// Auth for legit order
-		const takerOrderParamsSig = takerDriftClient.signSwiftOrderParamsMessage(
-			takerOrderParamsMessage
-		);
-
-		// Auth for non-legit order
-		const takerOrderParamsMessage2: SwiftOrderParamsMessage = {
-			swiftOrderParams: Object.assign({}, takerOrderParams, {
-				direction: PositionDirection.SHORT,
-				auctionStartPrice: new BN(34).mul(PRICE_PRECISION),
-				auctionEndPrice: new BN(33).mul(PRICE_PRECISION),
-				price: new BN(33).mul(PRICE_PRECISION),
-			}),
-			subAccountId: 0,
-			takeProfitOrderParams: null,
-			stopLossOrderParams: null,
-			uuid: Uint8Array.from(Buffer.from(nanoid(8))),
-			slot: new BN(await connection.getSlot()),
-		};
-
-		const takerOrderParamsSig2 = takerDriftClient.signSwiftOrderParamsMessage(
-			takerOrderParamsMessage2
-		);
-
-		const computeBudgetIx = ComputeBudgetProgram.setComputeUnitLimit({
-			units: 10_000_000,
+		const makerOrderParams = getLimitOrderParams({
+			marketIndex,
+			direction: PositionDirection.SHORT,
+			baseAssetAmount: BASE_PRECISION,
+			price: new BN(33).mul(PRICE_PRECISION),
+			userOrderId: 1,
+			postOnly: PostOnlyParams.MUST_POST_ONLY,
+			immediateOrCancel: true,
 		});
 
-		const ixs = await makerDriftClient.getPlaceSwiftTakerPerpOrderIxs(
-			takerDriftClient.encodeSwiftOrderParamsMessage(takerOrderParamsMessage),
-			takerOrderParamsSig,
-			0,
-			{
-				taker: await takerDriftClient.getUserAccountPublicKey(),
-				takerUserAccount: takerDriftClient.getUserAccount(),
-				takerStats: takerDriftClient.getUserStatsAccountPublicKey(),
-			}
+		const takerOrderParamsMessageEncoded =
+			takerDriftClient.encodeSwiftOrderParamsMessage(takerOrderParamsMessage);
+		const takerOrderParamsSig = takerDriftClient.signMessage(
+			takerOrderParamsMessageEncoded,
+			makerDriftClient.wallet.payer
 		);
 
-		const ixsForOrder2 = await makerDriftClient.getPlaceSwiftTakerPerpOrderIxs(
-			takerDriftClient.encodeSwiftOrderParamsMessage(takerOrderParamsMessage2),
-			takerOrderParamsSig2,
-			0,
-			{
-				taker: await takerDriftClient.getUserAccountPublicKey(),
-				takerUserAccount: takerDriftClient.getUserAccount(),
-				takerStats: takerDriftClient.getUserStatsAccountPublicKey(),
-			}
+		const ixs = [
+			ComputeBudgetProgram.setComputeUnitLimit({
+				units: 10_000_000,
+			}),
+		];
+		ixs.push(
+			...(await makerDriftClient.getPlaceAndMakeSwiftPerpOrderIxs(
+				takerDriftClient.encodeSwiftOrderParamsMessage(takerOrderParamsMessage),
+				takerOrderParamsSig,
+				uuid,
+				{
+					taker: await takerDriftClient.getUserAccountPublicKey(),
+					takerUserAccount: takerDriftClient.getUserAccount(),
+					takerStats: takerDriftClient.getUserStatsAccountPublicKey(),
+				},
+				makerOrderParams,
+				undefined,
+				undefined,
+				ixs
+			))
 		);
 
-		const badOrderTx = new Transaction();
-		badOrderTx.add(...[computeBudgetIx, ixs[0], ixsForOrder2[1]]);
 		try {
-			await makerDriftClient.sendTransaction(badOrderTx);
-			assert.fail('Should have failed');
+			const normalTx = new Transaction();
+			normalTx.add(...ixs);
+			await makerDriftClient.sendTransaction(normalTx);
+			assert.fail('should have thrown');
 		} catch (error) {
-			console.log((error as SendTransactionError).logs);
-			assert(
-				(error as SendTransactionError).logs.some((log) =>
-					log.includes('SigVerificationFailed')
-				)
-			);
-		}
-
-		const badSwiftTx = new Transaction();
-		badSwiftTx.add(...[computeBudgetIx, ixsForOrder2[0], ixs[1]]);
-		try {
-			await makerDriftClient.sendTransaction(badSwiftTx);
-			assert.fail('Should have failed');
-		} catch (error) {
-			console.log((error as SendTransactionError).logs);
-			assert(
-				(error as SendTransactionError).logs.some((log) =>
-					log.includes('SigVerificationFailed')
-				)
+			assert.equal(
+				error.transactionMessage,
+				'Transaction precompile verification failure InvalidAccountIndex'
 			);
 		}
 	});